For `_`, switch from Underscore to Lodash
Lodash will be easier to work with because it has more functions. It is also more compatible with advanced build tools.

Yes, this latest version of Lodash makes for a bigger download than the old version of Underscore I was using, but it’s still small. I don’t think the total download size is big enough to worry about right now. If it ever becomes worth worrying about, I can use a tree-shaking build utility to only include the code of the parts of Lodash I use, or import just the Lodash functions I need using per-method packages.