Rename and nest generateSkewFor… functions
The long names made it hard to tell the difference between these four functions. I think these new names, and the nesting so that functions are called very close to their definition, make it easier to follow.

I know that adding “NonSkipping” to the name of the last function “generatePotentialSkewForInaccurateNonSkippingNextTick” makes the name longer, but the name should have had that qualifier in the first place. The name “generatePotentialSkewForInaccurateNextTick” would imply the incorrect relationship between that function and “generateSkewForInaccurateNextTick”.

Moving these functions to their own file could also make them easier to read, but I want to keep the build system simple for now by not requiring compiling of the JS to have fast downloads.