Upgrade jQuery to v3.4.1
This upgrade is on the general principle of newer libraries being compatible with newer browsers. The uncompressed file size of the library increased by 5%, a negligible amount.

My JavaScript needs no changes to be compatible. My only calls are of the forms jQuery(selector), jQuery(callback), .prop(propertyName, value), and .on(events, handler), and these have not changed relevantly.